@@ -12,16 +12,16 @@ class MySQL57Driver extends MySQL56Driver
     protected function createColumnStructure(array $rawColumn): ColumnStructureInterface
     {
         $attribute = [];
-        if ((bool) preg_match('/auto_increment/', $rawColumn['EXTRA'])) {
+        if ((bool) preg_match('/auto_increment/', (string) $rawColumn['EXTRA'])) {
             $attribute[] = Attribute::AUTO_INCREMENT;
         }
         if ('YES' === $rawColumn['IS_NULLABLE']) {
             $attribute[] = Attribute::NULLABLE;
         }
-        if ((bool) preg_match('/unsigned/', $rawColumn['COLUMN_TYPE'])) {
+        if ((bool) preg_match('/unsigned/', (string) $rawColumn['COLUMN_TYPE'])) {
             $attribute[] = Attribute::UNSIGNED;
         }
-        if ((bool) preg_match('/STORED/', $rawColumn['EXTRA'])) {
+        if ((bool) preg_match('/STORED/', (string) $rawColumn['EXTRA'])) {
             $attribute[] = Attribute::STORED;
         }
 
